Reorganize heuristics
