+++ src/collection-facets/smart-facets/heuristics/browser-language/browser-language-heuristic.ts
@@ -1,4 +1,4 @@
-import type { SmartQueryHeuristic, SmartFacet } from '../models';
+import type { SmartQueryHeuristic, SmartFacet } from '../../models';
 
 export class BrowserLanguageHeuristic implements SmartQueryHeuristic {
   async getRecommendedFacets(): Promise<SmartFacet[]> {
