+++ src/collection-facets/smart-facets/smart-facet-heuristics.ts
@@ -1,19 +1,24 @@
 import { dedupe } from './dedupe';
-import { BrowserLanguageHeuristic } from './heuristics/browser-language-heuristic';
-import { QueryKeywordsHeuristic } from './heuristics/query-keywords-heuristic';
-import { WikidataHeuristic } from './heuristics/wikidata-heuristic';
+import { BrowserLanguageHeuristic } from './heuristics/browser-language/browser-language-heuristic';
+import { QueryKeywordsHeuristic } from './heuristics/query-keywords/query-keywords-heuristic';
+import { WikidataHeuristic } from './heuristics/wikidata/wikidata-heuristic';
 import type { SmartFacet, SmartQueryHeuristic } from './models';
 
 export class SmartQueryHeuristicGroup implements SmartQueryHeuristic {
-  private static readonly HEURISTICS = [
+  // Avoid collapsing the array onto one line
+  // prettier-ignore
+  private static readonly DEFAULT_HEURISTICS: (new () => SmartQueryHeuristic)[] = [
     QueryKeywordsHeuristic,
     WikidataHeuristic,
-    BrowserLanguageHeuristic,
+    BrowserLanguageHeuristic
   ];
 
-  async getRecommendedFacets(query: string): Promise<SmartFacet[]> {
-    const promises = SmartQueryHeuristicGroup.HEURISTICS.map(HeuristicCtor =>
-      new HeuristicCtor().getRecommendedFacets(query),
+  async getRecommendedFacets(
+    query: string,
+    heuristics = SmartQueryHeuristicGroup.DEFAULT_HEURISTICS
+  ): Promise<SmartFacet[]> {
+    const promises = heuristics.map(HeuristicCtor =>
+      new HeuristicCtor().getRecommendedFacets(query)
     );
 
     return dedupe((await Promise.all(promises)).flat());
